Fix bug in relative entropy derivative
Integrands in find_d_dx_D_PYgivenX_PY.m in both Amplitude Constrained
and Power Constrained folders were missing factor of ln(2) on the +1
term, as well as containing an extraneous factor of 1/2 overall. Due
to odd symmetry in the integral, this does not affect results of DAB.
However, it may cause issues if used to compute the Csiszar bound for
other applications.